EZP-31651: Removed logic from SearchController
