+++ src/lib/View/SearchListViewFilter.php
@@ -1,0 +1,120 @@
+<?php
+
+/**
+ * @copyright Copyright (C) eZ Systems AS. All rights reserved.
+ * @license For full copyright and license information view LICENSE file distributed with this source code.
+ */
+declare(strict_types=1);
+
+namespace Ibexa\Platform\Search\View;
+
+use eZ\Publish\API\Repository\ContentTypeService;
+use eZ\Publish\API\Repository\SectionService;
+use eZ\Publish\Core\MVC\ConfigResolverInterface;
+use eZ\Publish\Core\MVC\Symfony\View\Event\FilterViewBuilderParametersEvent;
+use eZ\Publish\Core\MVC\Symfony\View\ViewEvents;
+use Ibexa\Platform\Bundle\SearchBundle\Form\Data\SearchData;
+use Ibexa\Platform\Bundle\SearchBundle\Form\Type\SearchType;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\Form\FormFactoryInterface;
+use Symfony\Component\HttpFoundation\Request;
+
+class SearchListViewFilter implements EventSubscriberInterface
+{
+    /** @var \eZ\Publish\Core\MVC\ConfigResolverInterface */
+    private $configResolver;
+
+    /** @var \Symfony\Component\Form\FormFactoryInterface */
+    private $formFactory;
+
+    /** @var \eZ\Publish\API\Repository\SectionService */
+    private $sectionService;
+
+    /** @var \eZ\Publish\API\Repository\ContentTypeService */
+    private $contentTypeService;
+
+    public function __construct(
+        ConfigResolverInterface $configResolver,
+        FormFactoryInterface $formFactory,
+        SectionService $sectionService,
+        ContentTypeService $contentTypeService
+    ) {
+        $this->configResolver = $configResolver;
+        $this->formFactory = $formFactory;
+        $this->sectionService = $sectionService;
+        $this->contentTypeService = $contentTypeService;
+    }
+
+    public static function getSubscribedEvents()
+    {
+        return [ViewEvents::FILTER_BUILDER_PARAMETERS => 'handleSearchForm'];
+    }
+
+    /**
+     * @throws \Symfony\Component\OptionsResolver\Exception\UndefinedOptionsException
+     * @throws \Symfony\Component\OptionsResolver\Exception\OptionDefinitionException
+     * @throws \Symfony\Component\OptionsResolver\Exception\NoSuchOptionException
+     * @throws \Symfony\Component\OptionsResolver\Exception\MissingOptionsException
+     * @throws \Symfony\Component\OptionsResolver\Exception\AccessException
+     * @throws \Symfony\Component\OptionsResolver\Exception\InvalidOptionsException
+     * @throws \eZ\Publish\API\Repository\Exceptions\UnauthorizedException
+     * @throws \eZ\Publish\API\Repository\Exceptions\NotFoundException
+     */
+    public function handleSearchForm(FilterViewBuilderParametersEvent $event): void
+    {
+        $controllerAction = $event->getParameters()->get('_controller');
+
+        if (
+            'Ibexa\Platform\Bundle\SearchBundle\Controller\SearchController::searchAction' !== $controllerAction
+        ) {
+            return;
+        }
+
+        $request = $event->getRequest();
+
+        $search = $request->query->get('search');
+        $limit = $search['limit'] ? (int)$search['limit'] : $this->configResolver->getParameter('pagination.search_limit');
+        $page = $search['page'] ? (int)$search['page'] : 1;
+        $query = $search['query'] ?? '';
+        $section = null;
+        $creator = null;
+        $contentTypes = [];
+        $lastModified = $search['last_modified'] ?? [];
+        $created = $search['created'] ?? [];
+        $subtree = $search['subtree'] ?? null;
+        $searchLanguage = null;
+
+        if (!empty($search['section'])) {
+            $section = $this->sectionService->loadSection($search['section']);
+        }
+        if (!empty($search['content_types']) && \is_array($search['content_types'])) {
+            foreach ($search['content_types'] as $identifier) {
+                $contentTypes[] = $this->contentTypeService->loadContentTypeByIdentifier($identifier);
+            }
+        }
+
+        $form = $this->formFactory->create(
+            SearchType::class,
+            new SearchData(
+                $limit,
+                $page,
+                $query,
+                $section,
+                $contentTypes,
+                $lastModified,
+                $created,
+                $creator,
+                $subtree,
+                $searchLanguage
+            ),
+            [
+                'method' => Request::METHOD_GET,
+                'csrf_protection' => false,
+            ]
+        );
+
+        $event->getParameters()->add([
+            'form' => $form->handleRequest($request),
+        ]);
+    }
+}
